+++ apps/web/app/(root)/(admin)/admin/products/page.tsx
@@ -1,8 +1,8 @@
 "use client";
 
-import ModalAddProduct from "../../../../components/admin/modal/modal-product-add";
+import ModalAddProduct from "../../../../components/admin/components/modal/modal-product-add";
 import { TableProduct } from "../../../../components/admin/table/table-product";
-import Notification from "../../../../components/ui/Notification";
+import Notification from "../../../../components/admin/components/ui/Notification";
 import { useState } from "@repo/ui";
 const page = () => {
   return (
