+++ apps/web/app/auth/login/page.tsx
@@ -3,13 +3,15 @@ import Image from "next/image";
 // import styles from "./page.module.css";
 // import ButtonMain from "./components/Button";
 import { Container } from "@repo/ui";
-import LoginForm from "../../components/form/login-form";
+import { LoginForm } from "../../components/client/pages/LoginForm";
+// import LoginForm from "../../components/form/login-form";
 
 export default function Home() {
   return (
     <>
       <Container size="xl">
-        <LoginForm />
+        {/* <LoginForm /> */}
+        <LoginForm/>
       </Container>
     </>
   );
