+++ apps/web/app/components/admin/components/modal/modal-product-delete.tsx
@@ -15,7 +15,7 @@ import {
   useEffect,
   Notification,
 } from "@repo/ui";
-import request from "../../../utils/request";
+import request from "../../../../utils/request";
 
 interface ModalDeleteProductProps {
   productId: number;
