+++ apps/web/app/components/client/pages/Footer.tsx
@@ -1,0 +1,12 @@
+import { AiOutlineGithub, AiFillTwitterCircle } from "@repo/ui";
+import Link from "next/link";
+
+export const Footer = () => {
+  return (
+    <footer className="min-h-[100px] bg-beige mt-32 w-full  px-5 md:px-20 flex items-center justify-center rounded-t-2xl rounded-r-2xl">
+     
+        &copy;2023 This page is order food
+  
+    </footer>
+  );
+};
