+++ apps/web/app/components/client/pages/menu/header.module.css
@@ -1,0 +1,52 @@
+.mainSection {
+    padding-bottom: var(--mantine-spacing-sm);
+  }
+  
+  .userActive {
+    background-color: light-dark(
+      var(--mantine-color-white),
+      var(--mantine-color-dark-8)
+    );
+  }
+  
+  .tabsList {
+    &::before {
+      display: none;
+    }
+  }
+  
+  .tab {
+    font-weight: 500;
+    height: rem(38);
+    background-color: transparent;
+    position: relative;
+    bottom: -1px;
+  
+    &::before,
+    &::after {
+      background-color: light-dark(
+        var(--mantine-color-gray-2),
+        var(--mantine-color-dark-7)
+      ) !important;
+    }
+  
+    &:hover {
+      background-color: light-dark(
+        var(--mantine-color-gray-1),
+        var(--mantine-color-dark-5)
+      );
+    }
+  
+    &[data-active] {
+      background-color: light-dark(
+        var(--mantine-color-white),
+        var(--mantine-color-dark-7)
+      );
+      border-color: light-dark(
+        var(--mantine-color-gray-2),
+        var(--mantine-color-dark-7)
+      );
+      border-bottom-color: transparent;
+    }
+  }
+
