--- apps/web/app/components/user/navbar/navbar.module.css
@@ -1,67 +1,0 @@
-.navbar {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-between;
-  height: 100%;
-  padding: 0.2rem;
-  background-color: #f8f9fa;
-  /* width: rem(80px); */
-  /* height: rem(750px); */
-  padding: var(--mantine-spacing-md);
-  display: flex;
-  flex-direction: column;
-  border-right: rem(1px) solid
-    light-dark(var(--mantine-color-gray-3), var(--mantine-color-dark-4));
-}
-
-.navbarMain {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  gap: 1rem;
-}
-
-.link {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 20px;
-  /* width: 120px;
-  height: 120px; */
-  border-radius: 50%;
-  transition: background-color 0.3s;
-  /* width: rem(50px);
-    height: rem(50px); */
-  font-size: 20px;
-  border-radius: var(--mantine-radius-md);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: light-dark(var(--mantine-color-gray-7), var(--mantine-color-dark-0));
-
-  &:hover {
-    background-color: light-dark(
-      var(--mantine-color-gray-0),
-      var(--mantine-color-dark-5)
-    );
-  }
-
-  &[data-active] {
-    &,
-    &:hover {
-      background-color: var(--mantine-color-red-light);
-      color: var(--mantine-color-red-light-color);
-    }
-  }
-}
-.link:hover,
-.link[data-active] {
-  background-color: var(--mantine-color-red-light);
-  color: var(--mantine-color-red-light-color);
-}
-
-.active {
-  background-color: var(--mantine-color-red-light);
-}
