--- apps/web/app/components/user/order/order.module.css
@@ -1,16 +1,0 @@
-.navbar {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-between;
-  height: 100vh;
-  /* background-color: antiquewhite; */
-
-  display: flex;
-  flex-direction: column;
-  border-right: rem(1px) solid
-    light-dark(var(--mantine-color-gray-3), var(--mantine-color-dark-4));
-}
-.orderBox {
-  /* padding: var(--mantine-spacing-md); */
-}
