+++ packages/ui/src/index.ts
@@ -2,6 +2,7 @@ export * from "@mantine/notifications";
 export * from "@mantine/hooks";
 export * from "@mantine/core";
 export * from "react-icons/fa";
+export * from "react-icons/tb";
 export * from "react-icons/fc";
 export * from "react-icons/ai";
 export * from "react-icons/fi";
